Improve logging support
* remove log.py
* Adhere to common logging practise
