+++ tests/test_main.py
@@ -4,14 +4,13 @@ from docopt import DocoptExit
 import pytest
 import os
 
-from docstats.main import main
+from docstats.cli import main
 from unittest.mock import patch, Mock
 from configparser import ConfigParser, DuplicateSectionError, DuplicateOptionError
 
 
-@pytest.mark.skip
-def test_main():
-    """runs __main__.py"""
+def test_main(capsys):
+    """Checks, if __main__.py can be executed"""
     with pytest.raises(SystemExit):
         path = os.path.dirname(os.path.realpath(__file__)) + "/../src/docstats/__main__.py"
         exec(compile(open(path).read(), path, "exec"), {}, {"__name__": "__main__"})
@@ -27,7 +26,7 @@ def test_main_with_empty_args():
     DuplicateSectionError('fake-sec'),
     DuplicateOptionError('fake-sec', 'fake-opt'),
 ])
-@patch('docstats.main.parsecli')
+@patch('docstats.cli.parsecli')
 def test_main_with_ctrl_c(mock_parsecli, error):
     mock_parsecli.side_effect = error
     result = main([])
@@ -35,10 +34,10 @@ def test_main_with_ctrl_c(mock_parsecli, error):
 
 
 @patch('docstats.worker.work')
-@patch('docstats.main.os.makedirs')
-@patch('docstats.main.gettmpdir')
-@patch('docstats.main.parseconfig')
-@patch('docstats.main.parsecli')
+@patch('docstats.cli.os.makedirs')
+@patch('docstats.cli.gettmpdir')
+@patch('docstats.cli.parseconfig')
+@patch('docstats.cli.parsecli')
 def test_main_return_with_0(mock_parsecli, mock_parseconfig, mock_gettmpdir,
                             mock_makedirs, mock_work):
     def work(config, basedir, sections, jobs):
